Move ratings to separate module
