+++ src/app.module.ts
@@ -13,7 +13,7 @@ import { User } from "./users/entities/user.entity";
 import { Product } from "./products/entities/product.entity";
 import { Comment } from "./comments/comment.entity";
 import { Image } from "./images/image.entity";
-import { Rating } from "./products/entities/rating.entity";
+import { Rating } from "./ratings/rating.entity";
 import { Order } from "./orders/entities/order.entity";
 import { OrdersModule } from "./orders/orders.module";
 import { OrderRow } from "./orders/entities/order-row.entity";
